@@ -3,9 +3,9 @@ ext.versions = [
          * Build
          */
         minSdk            : 16,
-        compileSdk        : 27,
-        buildTools        : '27.0.3',
-        gradlePlugin      : '3.1.2',
+        compileSdk        : 28,
+        buildTools        : '28.0.3',
+        gradlePlugin      : '3.4.1',
 
         /**
          * App
