@@ -1,8 +1,9 @@
 package com.mensarb.onboarding;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.content.ContextCompat;
+
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
 
 import com.mensarb.onboarder.models.Page;
 import com.mensarb.onboarder.modules.OnboardingActivity;
